qgqx_api/qgqx/requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepdiff==6.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==3.0.2
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpath==0.82.2
MarkupSafe==2.1.5
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.2
pluggy==1.4.0
pycryptodome==3.20.0
PyMySQL==1.1.0
pytest==8.0.2
pytest-rerunfailures==13.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
urllib3==2.2.1
Werkzeug==3.0.1
zipp==3.17.0
